+++ app/tests/test_inverter_g3p.py
@@ -116,7 +116,7 @@ def test_method_calls(patch_conn_init, patch_conn_close):
     inverter.r_addr = ''
 
     spy1.assert_called_once()
-    spy1.assert_called_once_with(reader, writer, addr, None, client_mode=False)
+    spy1.assert_called_once_with(inverter, reader, writer, addr, None, client_mode=False)
 
     inverter.close()
     spy2.assert_called_once()
